feat: add AbstractOpa (#730)
* feat: add AbstractOpa

* feat: add AbstractOpa

* feat: add AbstractOpa

* feat: add AbstractOpa

* feat: add AbstractOpa

* fix: sonar deprecated
